table changed

+++ src/main/java/edu/segeyrozhkov/crud/app/service/UserService.java
@@ -8,9 +8,9 @@ public interface UserService {
 
     void updateUser(User user);
 
-    void removeUser(int id);
+    void removeUser(long id);
 
-    User getUserById(int id);
+    User getUserById(long id);
 
     List<User> userList();
 }
